// frontend/src/types/presignedUploadResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { PresignedURLResponse } from './presignedURLResponse';
import type { FileResponse } from './fileResponse';

export interface PresignedUploadResponse {
  presignedURL?: PresignedURLResponse;
  file?: FileResponse;
}